# openapi.yaml
openapi: 3.0.0
info:
  title: AIQuiz API
  description: Complete API documentation for the AIQuiz educational platform with AI-powered question generation and RAG-enabled document processing
  version: 1.0.0
  contact:
    name: AIQuiz Team
servers:
  - url: /api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/manager/auth/login
...
paths:
...
  /manager/subjects/{id}/topics/{topicId}/questions/download:
    get:
      summary: Download questions as file
      tags:
        - Manager Questions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, json, csv]
            default: pdf
        - name: verified
          in: query
          schema:
            type: string
            enum: [true, false, all]
            default: all
      responses:
        '200':
          description: Questions file download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # Questionnaire Management
...